copilot-shayer/requirements.txt
fastapi==0.115.12
openai==1.76.0
pydantic==2.11.3
PyPDF2==3.0.1
python-multipart==0.0.20
python-dotenv==1.1.0
uvicorn==0.34.2
